+++ test/specs/repo_client/extensions/scribunto-item.ts
@@ -4,6 +4,7 @@ import LoginPage from 'wdio-mediawiki/LoginPage.js';
 import { getTestString } from 'wdio-mediawiki/Util.js';
 import WikibaseApi from 'wdio-wikibase/wikibase.api.js';
 import ItemPage from '../../../helpers/pages/entity/item.page.js';
+import RepoClientLoginPage from '../../../helpers/pages/repo-client-login.page.js';
 import { utf8 } from '../../../helpers/read-file-encoding.js';
 import ExternalChange from '../../../types/external-change.js';
 
@@ -13,6 +14,13 @@ describe( 'Scribunto Item', function () {
 	let itemId: string;
 	const propertyValue = 'PropertyExampleStringValue';
 	const luaPageTitle = 'RepoClientLuaTest';
+
+	before( async function () {
+		await RepoClientLoginPage.login(
+			testEnv.vars.MW_ADMIN_NAME,
+			testEnv.vars.MW_ADMIN_PASS
+		);
+	} );
 
 	beforeEach( async function () {
 		await browser.skipIfExtensionNotPresent( this, 'Scribunto' );